correct g_associations_matrix
In line 92, if I understood correctly, the residuals are being divided by the denominator of the association index. In the paper (https://doi.org/10.1111/2041-210X.12383), Whitehead and James wrote: "When using a binomial model, they should be divided by the denominator original association index". However, note that this is wrong when models are built in R. In MATLAB, when running a binomial generalized linear model, the residuals are retrieved in the form of the response variable (i.e. the numerator). Then, dividing the residuals by the denominator is correct. This seems to be the reason why the authors suggest such division for binomial GLMs. Contrarily, the deviance residuals are returned by default. If so, the residuals must not be divided by the association index denominator in the assoc.gfi function.